Unknown misc updates.
The last commit to this project was nine years ago. This commit is for
whatever random uncommitted work was sitting around on my drive since
that last one. (It looks to be mostly around the laser-cut enclosure and
realtime clock, but I didn't inspect the details too closely.)

Because this project got moved to an archive on a network drive, the
file bits are a little messed up (smb shares show everything as +x). My
level of caring is low enough that I'm not going to try to correct that
right now.

I no longer have a great way of testing changes to this project.
Consider it abandonware.